Rename {RebrandBanner => AnnouncementBanner}
