+++ src/resources/app-en.json
@@ -229,6 +229,7 @@
   "alwaysAvailableWarning": "Once you set this event to always available, all scheduled events for this activity will be removed.",
   "any": "Any",
   "and": "and",
+  "announcementBanner": "<0>We are rebranding! </0> <1>Design updates are on the way — same great app, fresh new look. Curious? </1><2>Click to learn more.</2>",
   "applet_one": "Applet",
   "applet_other": "Applets",
   "appletActivities": "Applet activities",
@@ -1271,7 +1272,6 @@
   "quote": "Quote",
   "random": "Random",
   "rangeOfScores": "Possible Range of Scores",
-  "rebrandBanner": "<0>We are rebranding! </0> <1>Design updates are on the way — same great app, fresh new look. Curious? </1><2>Click to learn more.</2>",
   "record": "Record",
   "recordAudio": "Record <br /> Audio",
   "redo": "Redo",
