+++ src/resources/app-fr.json
@@ -229,6 +229,7 @@
   "alwaysAvailableWarning": "Si vous définissez cet événement comme toujours disponible, tous les événements programmés pour cette activité seront supprimés.",
   "any": "N'importe lequel",
   "and": "et",
+  "announcementBanner": "<0>Nous changeons de marque! </0> <1>Des mises à jour de design sont en cours — même application géniale, nouveau look frais. Curieux? </1><2>Cliquez pour en savoir plus.</2>",
   "applet_one": "Applet",
   "applet_other": "Applets",
   "appletActivities": "Activités de l'applet",
@@ -1270,7 +1271,6 @@
   "quote": "Citation",
   "random": "Aléatoire",
   "rangeOfScores": "Plage possible de scores",
-  "rebrandBanner": "<0>Nous changeons de marque! </0> <1>Des mises à jour de design sont en cours — même application géniale, nouveau look frais. Curieux? </1><2>Cliquez pour en savoir plus.</2>",
   "record": "Enregistrer",
   "recordAudio": "Enregistrer <br /> Audio",
   "redo": "Refaire",
