+++ src/shared/components/Banners/AnnouncementBanner/AnnouncementBanner.styles.ts
