+++ src/shared/components/Banners/AnnouncementBanner/AnnouncementBanner.test.tsx
@@ -7,9 +7,9 @@ import { mockedOwnerId } from 'shared/mock';
 import { getPreloadedState } from 'shared/tests/getPreloadedState';
 import { renderWithProviders } from 'shared/utils/renderWithProviders';
 
-import { GLOBAL_DISMISSED_KEY, RebrandBanner, getDismissedKey } from './RebrandBanner';
+import { GLOBAL_DISMISSED_KEY, AnnouncementBanner, getDismissedKey } from './AnnouncementBanner';
 
-const bannerTestId = 'rebrand-banner';
+const bannerTestId = 'announcement-banner';
 
 // Mock useLocation hook
 jest.mock('react-router-dom', () => ({
@@ -19,7 +19,7 @@ jest.mock('react-router-dom', () => ({
   })),
 }));
 
-describe('RebrandBanner', () => {
+describe('AnnouncementBanner', () => {
   // Get the mocked useLocation function
   const mockedUseLocation = useLocation as jest.Mock;
 
@@ -63,9 +63,9 @@ describe('RebrandBanner', () => {
     },
   };
 
-  describe('Rebrand Banner behavior for logged-in users', () => {
+  describe('Announcement Banner behavior for logged-in users', () => {
     test('shows banner when not dismissed', () => {
-      renderWithProviders(<RebrandBanner />, {
+      renderWithProviders(<AnnouncementBanner />, {
         preloadedState: {
           ...getPreloadedState(),
           ...mockAuthState,
@@ -76,7 +76,7 @@ describe('RebrandBanner', () => {
     });
 
     test('hides banner when dismissed', async () => {
-      renderWithProviders(<RebrandBanner />, {
+      renderWithProviders(<AnnouncementBanner />, {
         preloadedState: {
           ...getPreloadedState(),
           ...mockAuthState,
@@ -97,7 +97,7 @@ describe('RebrandBanner', () => {
       // Set user-level dismissal
       localStorage.setItem(getDismissedKey(mockedOwnerId), 'true');
 
-      renderWithProviders(<RebrandBanner />, {
+      renderWithProviders(<AnnouncementBanner />, {
         preloadedState: {
           ...getPreloadedState(),
           ...mockAuthState,
@@ -113,7 +113,7 @@ describe('RebrandBanner', () => {
         pathname: '/dashboard/some-applet-id',
       }));
 
-      renderWithProviders(<RebrandBanner />, {
+      renderWithProviders(<AnnouncementBanner />, {
         preloadedState: {
           ...getPreloadedState(),
           ...mockAuthState,
@@ -124,7 +124,7 @@ describe('RebrandBanner', () => {
     });
   });
 
-  describe('Rebrand Banner behavior on auth screen', () => {
+  describe('Announcment Banner behavior on auth screen', () => {
     beforeEach(() => {
       // Set location to auth route
       mockedUseLocation.mockImplementation(() => ({
@@ -133,7 +133,7 @@ describe('RebrandBanner', () => {
     });
 
     test('shows banner on auth screen when not dismissed', () => {
-      renderWithProviders(<RebrandBanner />, {
+      renderWithProviders(<AnnouncementBanner />, {
         preloadedState: {
           ...getPreloadedState(),
           ...mockAnonymousState,
@@ -144,7 +144,7 @@ describe('RebrandBanner', () => {
     });
 
     test('hides banner on auth screen when dismissed', async () => {
-      renderWithProviders(<RebrandBanner />, {
+      renderWithProviders(<AnnouncementBanner />, {
         preloadedState: {
           ...getPreloadedState(),
           ...mockAnonymousState,
@@ -165,7 +165,7 @@ describe('RebrandBanner', () => {
       // Set global dismissal
       localStorage.setItem(GLOBAL_DISMISSED_KEY, 'true');
 
-      renderWithProviders(<RebrandBanner />, {
+      renderWithProviders(<AnnouncementBanner />, {
         preloadedState: {
           ...getPreloadedState(),
           ...mockAnonymousState,
