+++ src/shared/components/Banners/AnnouncementBanner/index.ts
@@ -1,0 +1,1 @@
+export * from './AnnouncementBanner';
