--- src/shared/components/Banners/RebrandBanner/index.ts
@@ -1,1 +1,0 @@
-export * from './RebrandBanner';
